refactor: split homt.tsx components
